Drop in support for no_std float operations

+++ src/id.rs
@@ -8,15 +8,15 @@ pub(crate) type NodeId = usize;
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
 #[cfg_attr(not(any(feature = "std", feature = "alloc")), derive(hash32_derive::Hash32))]
-pub(crate) struct Id(u64);
+pub(crate) struct Id(usize);
 
 pub(crate) struct Allocator {
-    new_id: atomic::AtomicU64,
+    new_id: atomic::AtomicUsize,
 }
 
 impl Allocator {
     pub const fn new() -> Self {
-        Allocator { new_id: atomic::AtomicU64::new(0) }
+        Allocator { new_id: atomic::AtomicUsize::new(0) }
     }
 
     pub fn allocate(&self) -> Id {
